rename preloader component
